Moved Summary to its own source file
in the long run, I'd like to move every class to its own file as seems
customary in the C++ world.
